requirements.txt
requests
fpdf
Pillow
pytest
fpdf2